% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{hist.ipfraking}
\alias{hist.ipfraking}
\title{Histogram of raked weights}
\usage{
\method{hist}{ipfraking}(object, ...)
}
\description{
Implements generic \code{hist} method for output of \code{raking} function.
}
