% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{summary.ipfraking}
\alias{summary.ipfraking}
\title{Summary of Raked Weights}
\usage{
\method{summary}{ipfraking}(object, ...)
}
\description{
Implements generic \code{summary} method for output of \code{raking} function.
}
