% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raking-wt.r
\name{raking}
\alias{raking}
\title{Survey Raking in R}
\usage{
raking(
  formula,
  data,
  targets,
  namesTotals,
  maxiter = 50,
  epsilon = 1,
  freq = TRUE,
  freqWarn = 15,
  verbose = c("some", "none", "full"),
  trim = FALSE,
  trim.method = c("median", "mean"),
  trim.max = 6,
  trim.group = NULL
)
}
\arguments{
\item{formula}{a formula object, with the input/start weights on the left of the ~ operator and the raking variables on the right, with each raking variable separated by the + operator. Two-way variable raking can be done by using the * operator.}

\item{data}{data frame of respondents in which to interpret the variables named in the formula.}

\item{targets}{a single data frame or a list of data frames containing the target/calibration totals/figures. The raking variables listed in the formula must match the column names of those data frames. The target data frames must be listed in the same order as their corresponding raking variables in the formula.}

\item{namesTotals}{vector, where each element is the name of the column that contains the target/calibration totals in the corresponding data frame listed in targets.}

\item{maxiter}{(default 50) maximum number of iterations to attempt for convergence.}

\item{epsilon}{(default 1) convergence is declared if the maximum of the differences between marginal total of the raked weights and the target/calibration totals is less than or equal to epsilon. If epsilon<1 it is taken to be a fraction of the total of the sinput/start weights in the formula.}

\item{freq}{Logical. If true (default), frequency tables are displayed (this is useful to see if cells should be collapsed).}

\item{freqWarn}{a warning message will be displayed if one or more cells contained fewer than this number (default 15) of observations.}

\item{verbose}{controls the amount of output displayed on the screen.}

\item{trim}{logical, if true large weights will be trimmed. This is done using a technique very similar to the IGCV method of Izrael et al. (2009)}

\item{trim.method}{"median" (default) or "mean"; see trim.max}

\item{trim.max}{(default 6) each weight w is trimmed if w > trim.max * trim.method(weights)}

\item{trim.group}{if NULL (default), the overall mean/median of the weights is used when trimming; if trim.group is the name of a column of data, then the weights will be trimmed based on which group of trim.group they belong to.}
}
\description{
Performs raking of survey weights, using a technique very similar to the SAS IHB Macro
}
\examples{
   raking(wts.start ~ A + B*C, data.dat, list(totalsA.dat, totalsBC.dat), "nb.users")
   raking(wts.start ~ userGrp*ageGrp + userGrp*sex + ethnic*userGrp + region, tmpUS.dat, list(targetsUS_ew1_age.dat,
      targetsUS_ew1_sex.dat, targetsUS_ew1_ethnic.dat, targetsUS_ew1_region.dat2), "nb.users", freq=F, verbose="some")
}
\references{
\itemize{
\item Izrael, Hoaglin & Battaglia (2000), \href{http://www2.sas.com/proceedings/sugi25/25/st/25p258.pdf}{"A SAS Macro for Balancing a Weighted Sample"}, Proceedings of the 25th Annual SAS Users Group International Conference, Paper 258
\item Izrael, Battaglia & Frankel (2009), \href{https://support.sas.com/resources/papers/proceedings09/247-2009.pdf}{"Extreme Survey Weight Adjustment as a Component of Sample Balancing (a.k.a. Raking)"}, SAS Global Forum, Paper 247-2009
}
}
